+++ plugins/woocommerce-admin/tests/reports/class-wc-tests-reports-revenue-stats.php
@@ -38,7 +38,7 @@ class WC_Admin_Tests_Reports_Revenue_Stats extends WC_Unit_Test_Case {
 		$order->save();
 
 		// /reports/revenue/stats is mapped to Orders_Data_Store.
-		$data_store = new WC_Admin_Reports_Orders_Data_Store();
+		$data_store = new WC_Admin_Reports_Orders_Stats_Data_Store();
 
 		$start_time = date( 'Y-m-d H:00:00', $order->get_date_created()->getOffsetTimestamp() );
 		$end_time   = date( 'Y-m-d H:59:59', $order->get_date_created()->getOffsetTimestamp() );